requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.1
pydantic>=2.6.1
scikit-learn>=1.3.0
nltk>=3.8.1
textblob>=0.17.1
